Fix (TeamMate) int warning

+++ ConsoleApplication1/Battle/TeamMate.cpp
@@ -198,13 +198,13 @@ int TeamMate::reload(const bool &healing)
 	for (unsigned int i(0); i < m_characteristics.size(); i++)
 	{
 		//m_characteristics.at(i).setMaxPoints(m_courbs.at(i).result(m_level));
-		m_characteristics.at(i).setPoints(m_courbs.at(i).result(m_level));
+		m_characteristics.at(i).setPoints((int)m_courbs.at(i).result(m_level));
 	}
 
-	m_health.setMaxPoints(m_healthCourb.result(m_level));
+	m_health.setMaxPoints((int)m_healthCourb.result(m_level));
 	if (healing)
 		m_health.setPoints(m_health.getMaxPoints());
-	m_mana.setMaxPoints(m_manaCourb.result(m_level));
+	m_mana.setMaxPoints((int)m_manaCourb.result(m_level));
 	if (healing)
 		m_mana.setPoints(m_mana.getMaxPoints());
